Clean up GenericBsa and C wrapper code
Tests on Linux are failing for some unknown reason, but they pass
on Windows, so take the opportunity to generally improve code
quality, and hope that somehow fixes the issue.

* Improve const-correctness
* Make members static where applicable
* Reduce code repetition
* Use STL algorithms to simplify operations
* Combine adjacent try-catch blocks targeting different exceptions

One point worth mentioning is that the GenericBsa::Extract()
function that operates on multiple assets now opens up a new input
file stream for each asset, which is less efficient, but this is
just because it's using the single-file Extract(), and an overload
with an existing stream could be added if it's found to be a
significant effect.